Unbloat GenericObject by splitting it into a factory
