+++ GDAPI/GDAPI/Objects/GeometryDash/LevelObjects/GeneralObject.cs
@@ -26,30 +26,6 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
     /// <summary>Represents a general object.</summary>
     public class GeneralObject
     {
-        private static Type[] objectTypes;
-        private static ObjectTypeInfoDictionary objectTypeDictionary;
-        private static Dictionary<int, Type> propertyTypeInfo;
-
-        static GeneralObject()
-        {
-            // Get object property types from the attributes that are assigned to the enum fields
-            var fields = typeof(ObjectProperty).GetFields();
-            // default(ObjectParameter) is necessary otherwise an exception will be thrown
-            // Reflection really sucks dick when it comes to enums
-            var info = fields.Select(i => new KeyValuePair<int, Type>((int)i.GetValue(default(ObjectProperty)), i.GetCustomAttribute<ObjectPropertyTypeAttribute>()?.Type));
-            propertyTypeInfo = new Dictionary<int, Type>();
-            foreach (var i in info)
-                propertyTypeInfo.TryAdd(i.Key, i.Value);
-            // I added those 5 lines of code only to realize that I actually do not necessarily need them, hopefully they'll turn out any useful in the near future:tm:
-            // For the time being there is no reason to change them at all
-
-            objectTypes = typeof(GeneralObject).Assembly.GetTypes().Where(t => typeof(GeneralObject).IsAssignableFrom(t)).ToArray();
-            objectTypeDictionary = new ObjectTypeInfoDictionary();
-            foreach (var t in objectTypes.Select(t => ObjectTypeInfo.GetInfo(t)))
-                if (t != null)
-                    objectTypeDictionary.Add((IFirstWideDoubleKeyedObject<int?, Type>)t);
-        }
-
         private short[] groupIDs = new short[0]; // Create a ComparableArray<T> class and use it instead
         private BitArray16 bools = new BitArray16();
         private short objectID, el1, el2, zLayer, zOrder, color1ID, color2ID;
@@ -540,7 +516,7 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
         public bool IsWithinRange(double startingX, double startingY, double endingX, double endingY) => startingX <= X && endingX >= X && startingY <= Y && endingY >= Y;
 
         /// <summary>Retrieves the properties of this object's type.</summary>
-        public PropertyAccessInfoDictionary GetObjectProperties() => objectTypeDictionary[GetType()]?.Properties as PropertyAccessInfoDictionary;
+        public PropertyAccessInfoDictionary GetObjectProperties() => LevelObjectFactory.GetObjectTypeInfo(GetType())?.Properties as PropertyAccessInfoDictionary;
 
         /// <summary>Returns the name of the object type.</summary>
         /// <param name="lowerLastWord">Determines whether the last word will be converted to lower case.</param>
@@ -585,91 +561,6 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
         public static bool operator ==(GeneralObject left, GeneralObject right) => left.Equals(right);
         public static bool operator !=(GeneralObject left, GeneralObject right) => !left.Equals(right);
 
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(int objectID)
-        {
-            objectTypeDictionary.TryGetValue(objectID, out var t);
-            if (t is ObjectTypeInfo i)
-            {
-                if (i.NonGeneratableAttribute != null)
-                    throw new InvalidOperationException(i.NonGeneratableAttribute.ExceptionMessage);
-                var instance = i.Constructor.Invoke(null) as GeneralObject;
-                instance.objectID = (short)objectID; // Ensure the new object has the specified ID in case of a generic case
-                return instance;
-            }
-
-            return new GeneralObject(objectID);
-        }
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(TriggerType objectID) => GetNewObjectInstance((int)objectID);
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(SpecialObjectType objectID) => GetNewObjectInstance((int)objectID);
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(SpecialBlockType objectID) => GetNewObjectInstance((int)objectID);
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(OrbType objectID) => GetNewObjectInstance((int)objectID);
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(PadType objectID) => GetNewObjectInstance((int)objectID);
-        /// <summary>Returns a new instance of the appropriate class of an object based on its object ID.</summary>
-        /// <param name="objectID">The object ID of the new object.</param>
-        public static GeneralObject GetNewObjectInstance(PortalType objectID) => GetNewObjectInstance((int)objectID);
-
-        /// <summary>Returns the common properties found in the specified <seealso cref="LevelObjectCollection"/>.</summary>
-        /// <param name="collection">The collection whose common object properties will be evaluated and returned.</param>
-        public static PropertyAccessInfoDictionary GetCommonProperties(IEnumerable<GeneralObject> collection) => GetCommonProperties(collection, null);
-        /// <summary>Returns the common properties found in the specified <seealso cref="LevelObjectCollection"/> from a starting list of common properties.</summary>
-        /// <param name="collection">The collection whose common object properties will be evaluated and returned.</param>
-        /// <param name="startingDictionary">The starting dictionary of common properties that will be merged with the resulting list.</param>
-        public static PropertyAccessInfoDictionary GetCommonProperties(IEnumerable<GeneralObject> collection, PropertyAccessInfoDictionary startingDictionary)
-        {
-            var objectTypes = GetCollectionObjectTypeInfo(collection);
-
-            KeyedPropertyInfoDictionary<int?> lockedDictionary;
-            PropertyAccessInfoDictionary result;
-            if (collection != null)
-                result = new PropertyAccessInfoDictionary(lockedDictionary = startingDictionary);
-            else
-                result = new PropertyAccessInfoDictionary(lockedDictionary = objectTypes.First().Properties);
-
-            foreach (var t in objectTypes)
-                foreach (var p in lockedDictionary)
-                    if (!t.Properties.ContainsKey(p.Key))
-                        result.Remove(p.Key);
-
-            return result;
-        }
-        /// <summary>Returns all the available object properties found in the specified <seealso cref="LevelObjectCollection"/>.</summary>
-        /// <param name="collection">The collection whose all available object properties will be evaluated and returned.</param>
-        public static PropertyAccessInfoDictionary GetAllAvailableProperties(IEnumerable<GeneralObject> collection) => GetAllAvailableProperties(collection, null);
-        /// <summary>Returns all the available object properties found in the specified <seealso cref="LevelObjectCollection"/> from a starting hash set of available properties.</summary>
-        /// <param name="collection">The collection whose all available object properties will be evaluated and returned.</param>
-        /// <param name="startingDictionary">The starting dictionary of available properties that will be merged with the resulting hash set.</param>
-        public static PropertyAccessInfoDictionary GetAllAvailableProperties(IEnumerable<GeneralObject> collection, PropertyAccessInfoDictionary startingDictionary)
-        {
-            var objectTypes = GetCollectionObjectTypeInfo(collection);
-
-            PropertyAccessInfoDictionary result;
-            if (collection != null)
-                result = new PropertyAccessInfoDictionary(startingDictionary);
-            else
-                result = new PropertyAccessInfoDictionary();
-
-            foreach (var t in objectTypes)
-                foreach (var p in t.Properties)
-                    result.Add(p.Value as PropertyAccessInfo);
-
-            return result;
-        }
-        private static HashSet<ObjectTypeInfo> GetCollectionObjectTypeInfo(IEnumerable<GeneralObject> collection)
-        {
-            return new HashSet<ObjectTypeInfo>(collection.Select(o => objectTypeDictionary[o.GetType()]));
-        }
 
         /// <summary>Attempts to get the common value of an object property from a collection of objects given its ID.</summary>
         /// <typeparam name="T">The value type of the property.</typeparam>
@@ -743,6 +634,14 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
             return property;
         }
 
+        public PropertyAccessInfo GetPropertyAccessInfo(int ID)
+        {
+            KeyedPropertyInfo<int?> p = null;
+            LevelObjectFactory.TryGetObjectTypeInfo(GetType(), out var value);
+            value?.Properties.TryGetValue(ID, out p);
+            return p as PropertyAccessInfo;
+        }
+
         public override string ToString()
         {
             var s = new StringBuilder();
@@ -761,13 +660,6 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
             return s.ToString();
         }
 
-        private PropertyAccessInfo GetPropertyAccessInfo(int ID)
-        {
-            KeyedPropertyInfo<int?> p = null;
-            objectTypeDictionary.TryGetValue(GetType(), out var value);
-            value?.Properties.TryGetValue(ID, out p);
-            return p as PropertyAccessInfo;
-        }
         // I swear to fucking goodness I made this at 4:30 AM, this will be redone somewhere else it's just a temporary fix so that we release please have mercy
         // During the writing of this function a lot of WHEEZEs dropped
         private string GetAppropriateStringRepresentation(object thing)
@@ -788,149 +680,6 @@ namespace GDAPI.Objects.GeometryDash.LevelObjects
                 // Please tell me there are no more things that break
             }
             return thing.ToString();
-        }
-
-        public class ObjectTypeInfoDictionary : CachedTypeInfoDictionary<int?, int?>
-        {
-            public new ObjectTypeInfo this[int? key] => base[key] as ObjectTypeInfo;
-            public new ObjectTypeInfo this[Type key] => base[key] as ObjectTypeInfo;
-        }
-        public class ObjectTypeInfo : FirstWideCachedTypeInfo<int?, int?>
-        {
-            private static Func<Type, Type, PropertyInfo, PropertyAccessInfo> GetAppropriateGenericA;
-            private static Func<Type, PropertyInfo, PropertyAccessInfo> GetAppropriateGenericB;
-
-            public NonGeneratableAttribute NonGeneratableAttribute { get; }
-
-            public int?[] ObjectIDs { get; protected set; }
-
-            public override int?[] Key1 => ObjectIDs;
-            public override int? ConvertedKey => Key1 != null && Key1.Length > 0 ? Key1[0] : 0;
-
-            static ObjectTypeInfo()
-            {
-                //GenerateSelfExecutingCode();
-                // Uncomment this when System.CodeDom is referred
-            }
-
-            public ObjectTypeInfo(Type objectType)
-                : base(objectType)
-            {
-                NonGeneratableAttribute = objectType.GetCustomAttribute<NonGeneratableAttribute>();
-                InitializeObjectIDs();
-            }
-
-            protected virtual void InitializeObjectIDs()
-            {
-                ObjectIDs = ObjectType.GetCustomAttribute<ObjectIDsAttribute>()?.ObjectIDs.Cast<int?>().ToArray() ?? Array.Empty<int?>();
-                if (typeof(SpecialObject).IsAssignableFrom(ObjectType) && !ObjectType.IsAbstract)
-                {
-                    var validObjectIDs = (int[])ObjectTypeProperties.Where(p => p.Name == "ValidObjectIDs").First().GetValue(Constructor.Invoke(null));
-                    if (validObjectIDs != null)
-                        ObjectIDs = validObjectIDs.Cast<int?>().ToArray();
-                }
-            }
-
-            protected sealed override KeyedPropertyInfo<int?> CreateProperty(PropertyInfo p) => new PropertyAccessInfo(p);
-            // TODO: Use the following line instead
-            //protected sealed override PropertyAccessInfo CreateProperty(PropertyInfo p) => GetAppropriateGenericA?.Invoke(p.DeclaringType, p.PropertyType, p);
-
-            protected virtual string GetValidObjectIDs()
-            {
-                var s = new StringBuilder();
-                for (int i = 0; i < ObjectIDs.Length; i++)
-                    s.Append($"{ObjectIDs[i]}, ");
-                s.Remove(s.Length - 2, 2);
-                return s.ToString();
-            }
-
-            public static ObjectTypeInfo GetInfo(Type objectType)
-            {
-                if (objectType.GetCustomAttribute<ObjectIDAttribute>() != null)
-                    return new SingleObjectIDTypeInfo(objectType);
-                return new ObjectTypeInfo(objectType);
-            }
-
-            private static void GenerateSelfExecutingCode()
-            {
-                const string className = "PropertyAccessInfoGenerator";
-                const string methodAName = "GetAppropriateGenericA";
-                const string methodBName = "GetAppropriateGenericB";
-
-                var getAppropriateGenericACode = new StringBuilder();
-                var getAppropriateGenericBCode = new StringBuilder();
-
-                // With the current indentation, the produced source is badly formatted, but at least this looks better
-                // In the following code, nameof and constants have been purposefully selectively used for "obvious" reasons
-                var propertyTypes = new HashSet<Type>();
-                foreach (var o in objectTypes)
-                {
-                    var types = o.GetProperties().Select(p => p.PropertyType);
-                    foreach (var t in types)
-                        propertyTypes.Add(t);
-                    var fullName = o.FullName;
-                    getAppropriateGenericACode.Append($@"
-                    if (objectType == typeof({fullName}))
-                        return {methodBName}<{fullName}>(propertyType, info);");
-                }
-                foreach (var p in propertyTypes)
-                {
-                    var fullName = p.FullName;
-                    getAppropriateGenericBCode.Append($@"
-                    if (propertyType == typeof({fullName}))
-                        return new {nameof(PropertyAccessInfo)}<TO, {fullName}>(info);");
-                }
-
-                var allCode =
-$@"
-using System;
-using System.Reflection;
-
-namespace GDAPI.Special
-{{
-    public static partial class {className}
-    {{
-        public static {nameof(PropertyAccessInfo)} {methodAName}(Type objectType, Type propertyType, PropertyInfo info)
-        {{
-            {getAppropriateGenericACode}
-            throw new Exception(); // Don't make compiler sad
-        }}
-        public static {nameof(PropertyAccessInfo)} {methodBName}<TO>(Type propertyType, PropertyInfo info)
-        {{
-            {getAppropriateGenericBCode}
-            throw new Exception();
-        }}
-    }}
-}}
-";
-                // Compile all code
-                var provider = new CSharpCodeProvider();
-                var options = new CompilerParameters(new[] { "mscorlib.dll", "System.Core.dll", "System.dll", "System.CodeDom.dll" });
-                var assembly = provider.CompileAssemblyFromSource(options, allCode).CompiledAssembly;
-                var type = assembly.GetType(className);
-                GetAppropriateGenericA = type.GetMethod(methodAName).CreateDelegate(typeof(Func<Type, Type, PropertyInfo, PropertyAccessInfo>)) as Func<Type, Type, PropertyInfo, PropertyAccessInfo>;
-                GetAppropriateGenericB = type.GetMethod(methodBName).CreateDelegate(typeof(Func<Type, PropertyInfo, PropertyAccessInfo>)) as Func<Type, PropertyInfo, PropertyAccessInfo>; // maybe useless?
-            }
-
-            public override string ToString() => $"{GetValidObjectIDs()} - {ObjectType.Name}";
-        }
-        public class SingleObjectIDTypeInfo : ObjectTypeInfo
-        {
-            public int? ObjectID
-            {
-                get => ObjectIDs[0];
-                private set => ObjectIDs[0] = value;
-            }
-
-            public SingleObjectIDTypeInfo(Type objectType) : base(objectType) { }
-
-            protected override void InitializeObjectIDs()
-            {
-                ObjectIDs = new int?[1];
-                ObjectID = ObjectType.GetCustomAttribute<ObjectIDAttribute>()?.ObjectID;
-            }
-
-            protected override string GetValidObjectIDs() => ObjectID.ToString();
         }
     }
 }
